+++ src/frontend/apps/impress/src/features/docs/doc-header/components/DocToolBox.tsx
@@ -4,11 +4,19 @@ import {
   useToastProvider,
 } from '@openfun/cunningham-react';
 import { useRouter } from 'next/router';
-import React, { useState } from 'react';
+import { useState } from 'react';
 import { useTranslation } from 'react-i18next';
+import { css } from 'styled-components';
 
-import { Box, DropButton, IconOptions } from '@/components';
+import {
+  Box,
+  DropdownMenu,
+  DropdownMenuOption,
+  Icon,
+  IconOptions,
+} from '@/components';
 import { useAuthStore } from '@/core';
+import { useCunninghamTheme } from '@/cunningham';
 import {
   useEditorStore,
   usePanelEditorStore,
@@ -32,16 +40,81 @@ export const DocToolBox = ({ doc }: DocToolBoxProps) => {
     query: { versionId },
   } = useRouter();
   const { t } = useTranslation();
+  const { spacingsTokens, colorsTokens } = useCunninghamTheme();
+
+  const spacings = spacingsTokens();
+  const colors = colorsTokens();
+
   const [isModalShareOpen, setIsModalShareOpen] = useState(false);
   const [isModalRemoveOpen, setIsModalRemoveOpen] = useState(false);
   const [isModalPDFOpen, setIsModalPDFOpen] = useState(false);
-  const [isDropOpen, setIsDropOpen] = useState(false);
+
   const { setIsPanelOpen, setIsPanelTableContentOpen } = usePanelEditorStore();
   const [isModalVersionOpen, setIsModalVersionOpen] = useState(false);
   const { isSmallMobile } = useResponsiveStore();
   const { authenticated } = useAuthStore();
   const { editor } = useEditorStore();
   const { toast } = useToastProvider();
+
+  const options: DropdownMenuOption[] = [
+    ...(isSmallMobile
+      ? [
+          {
+            label: t('Share'),
+            icon: 'upload',
+            callback: () => {
+              setIsModalShareOpen(true);
+            },
+          },
+          {
+            label: t('Export'),
+            icon: 'download',
+            callback: () => {
+              setIsModalPDFOpen(true);
+            },
+          },
+        ]
+      : []),
+    {
+      label: t('Version history'),
+      icon: 'history',
+      disabled: !doc.abilities.versions_list,
+      callback: () => {
+        setIsPanelOpen(true);
+        setIsPanelTableContentOpen(false);
+      },
+    },
+    {
+      label: t('Table of contents'),
+      icon: 'summarize',
+      callback: () => {
+        setIsPanelOpen(true);
+        setIsPanelTableContentOpen(true);
+      },
+    },
+    {
+      label: t('Copy as {{format}}', { format: 'Markdown' }),
+      icon: 'content_copy',
+      callback: () => {
+        void copyCurrentEditorToClipboard('markdown');
+      },
+    },
+    {
+      label: t('Copy as {{format}}', { format: 'HTML' }),
+      icon: 'content_copy',
+      callback: () => {
+        void copyCurrentEditorToClipboard('html');
+      },
+    },
+    {
+      label: t('Delete document'),
+      icon: 'delete',
+      disabled: !doc.abilities.destroy,
+      callback: () => {
+        setIsModalRemoveOpen(true);
+      },
+    },
+  ];
 
   const copyCurrentEditorToClipboard = async (
     asFormat: 'html' | 'markdown',
@@ -87,9 +160,10 @@ export const DocToolBox = ({ doc }: DocToolBoxProps) => {
           </Button>
         </Box>
       )}
-      <Box $direction="row" $margin={{ left: 'auto' }} $gap="1rem">
-        {authenticated && (
+      <Box $direction="row" $margin={{ left: 'auto' }} $gap={spacings['2xs']}>
+        {authenticated && !isSmallMobile && (
           <Button
+            color="primary-text"
             onClick={() => {
               setIsModalShareOpen(true);
             }}
@@ -98,91 +172,34 @@ export const DocToolBox = ({ doc }: DocToolBoxProps) => {
             {t('Share')}
           </Button>
         )}
-        <DropButton
-          button={
-            <IconOptions
-              isOpen={isDropOpen}
-              aria-label={t('Open the document options')}
-            />
-          }
-          onOpenChange={(isOpen) => setIsDropOpen(isOpen)}
-          isOpen={isDropOpen}
-        >
-          <Box>
-            {doc.abilities.versions_list && (
-              <Button
-                onClick={() => {
-                  setIsPanelOpen(true);
-                  setIsPanelTableContentOpen(false);
-                  setIsDropOpen(false);
-                }}
-                color="primary-text"
-                icon={<span className="material-icons">history</span>}
-                size="small"
-              >
-                {t('Version history')}
-              </Button>
-            )}
-            <Button
-              onClick={() => {
-                setIsPanelOpen(true);
-                setIsPanelTableContentOpen(true);
-                setIsDropOpen(false);
-              }}
-              color="primary-text"
-              icon={<span className="material-icons">summarize</span>}
-              size="small"
-            >
-              {t('Table of contents')}
-            </Button>
-            <Button
-              onClick={() => {
-                setIsModalPDFOpen(true);
-                setIsDropOpen(false);
-              }}
-              color="primary-text"
-              icon={<span className="material-icons">file_download</span>}
-              size="small"
-            >
-              {t('Export')}
-            </Button>
-            {doc.abilities.destroy && (
-              <Button
-                onClick={() => {
-                  setIsModalRemoveOpen(true);
-                  setIsDropOpen(false);
-                }}
-                color="primary-text"
-                icon={<span className="material-icons">delete</span>}
-                size="small"
-              >
-                {t('Delete document')}
-              </Button>
-            )}
-            <Button
-              onClick={() => {
-                setIsDropOpen(false);
-                void copyCurrentEditorToClipboard('markdown');
-              }}
-              color="primary-text"
-              icon={<span className="material-icons">content_copy</span>}
-              size="small"
-            >
-              {t('Copy as {{format}}', { format: 'Markdown' })}
-            </Button>
-            <Button
-              onClick={() => {
-                setIsDropOpen(false);
-                void copyCurrentEditorToClipboard('html');
-              }}
-              color="primary-text"
-              icon={<span className="material-icons">content_copy</span>}
-              size="small"
-            >
-              {t('Copy as {{format}}', { format: 'HTML' })}
-            </Button>
-          </Box>
-        </DropButton>
+        {!isSmallMobile && (
+          <Button
+            color="primary-text"
+            icon={
+              <Icon iconName="download" $theme="primary" $variation="800" />
+            }
+            onClick={() => {
+              setIsModalPDFOpen(true);
+            }}
+            size={isSmallMobile ? 'small' : 'medium'}
+          />
+        )}
+        <DropdownMenu options={options}>
+          <IconOptions
+            isHorizontal
+            $theme="primary"
+            $radius={spacings['3xs']}
+            $css={
+              isSmallMobile
+                ? css`
+                    padding: 10px;
+                    border: 1px solid ${colors['greyscale-300']};
+                  `
+                : ''
+            }
+            aria-label={t('Open the document options')}
+          />
+        </DropdownMenu>
       </Box>
       {isModalShareOpen && (
         <ModalShare onClose={() => setIsModalShareOpen(false)} doc={doc} />